moved 'get_main_window' to nuke lib
